requirements.txt
aniso8601==7.0.0
cached-property==1.5.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
config==0.4.2
Django==2.2.10
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-graphql-jwt==0.3.1
django-url-filter==0.3.15
djangorestframework==3.11.0
docopt==0.6.2
enum-compat==0.0.3
google-auth==1.11.3
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.9
Jinja2==2.11.1
kubernetes==11.0.0
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
uWSGI==2.0.18
websocket-client==0.57.0
